requirements.txt
Flask
Werkzeug
Flask-WTF
WTForms
python-dotenv
pandas
numpy
PyMySQL
mysql-connector-python
Flask-MySQLdb
scikit-learn
xgboost
catboost
joblib